requirements.txt
gunicorn
parse~=1.19.0
pytest
mysql-connector-python~=8.0.32
WebOb~=1.8.7